laid out front page and did some more styling work on blog page.
going to have to start implementing the admin section soon to create blog posts,

and probably around the same time work on the web backend to have the api end points for calls.

much doco needed to be organised. (maybe just do entire front end and write a requirements list up, and decide on mock endpoints then plug them in as i go.