setup balancer and webfront to work,
fixed up bat files to boot up and close down the docker containers for testing purposes. (also connects them all to the same internal network. ).

todo: Make backend api calls that mess with a google cloud DB, and also setup a way to boot the emulator for it.
todo: Make frontend call the backend to get and set data